カット時間の指定機能を追加しました。.gitignoreを変更しました

+++ Form3.cs
@@ -26,6 +26,10 @@ namespace プロコン部チーム_0622_TEST
         private void Form3_Load(object sender, EventArgs e)
         {
             axWindowsMediaPlayer1.uiMode = "none";
+            Properties.Settings.Default.cut_start_time = "00:00:00.00"; //カット開始時間の初期値
+            Properties.Settings.Default.cut_end_time = "00:00:00.00";   //カット終了時間の初期値
+
+            axWindowsMediaPlayer1.URL = Properties.Settings.Default.raw_movie_filepath; //settingsファイルから動画のパスを読み込み
 
         }
 
@@ -33,8 +37,8 @@ namespace プロコン部チーム_0622_TEST
         {
             string input_filepath = Properties.Settings.Default.raw_movie_filepath;     //出力ファイルパスを.settingsファイルから読み込み
             string output_filepath = Path.Combine(Properties.Settings.Default.folderpath, "cut_output.mp4"); //カット後のファイルの保存先パス
-            string start_time = "00:00:10";                             //カットする範囲の開始地点の指定変数
-            string end_time = "00:10:00";                               //カットする範囲の終了地点の指定変数
+            string start_time = Properties.Settings.Default.cut_start_time;                           //カットする範囲の開始地点の指定変数
+            string end_time = Properties.Settings.Default.cut_end_time;                               //カットする範囲の終了地点の指定変数
 
 
             //カット処理
@@ -55,9 +59,14 @@ namespace プロコン部チーム_0622_TEST
             }
 
             string input_filename = Path.GetFileName(input_filepath);       //元のファイル名を取得
-            File.Delete(input_filepath); //元のファイルを削除
 
-            File.Move(output_filepath, Path.Combine(Properties.Settings.Default.folderpath, input_filename));//カット後のファイルを元のファイル名に変更
+            if (output_filepath == input_filepath)
+            {
+                File.Delete(input_filepath); //元のファイルを削除
+
+                File.Move(output_filepath, input_filename);//カット後のファイルを元のファイル名に変更
+            }
+
 
 
 
@@ -95,6 +104,7 @@ namespace プロコン部チーム_0622_TEST
 
 
                     axWindowsMediaPlayer1.URL = select_mp4_filepath;
+                    Properties.Settings.Default.raw_movie_filepath = select_mp4_filepath;
                 }
                 else
                 {
@@ -145,7 +155,7 @@ namespace プロコン部チーム_0622_TEST
                 // 時間表示ラベルを更新
                 TimeSpan current = TimeSpan.FromSeconds(axWindowsMediaPlayer1.Ctlcontrols.currentPosition);
                 TimeSpan total = TimeSpan.FromSeconds(axWindowsMediaPlayer1.currentMedia.duration);
-                time_display_label.Text = $"{current.ToString(@"hh\:mm\:ss\.ff")} / {total.ToString(@"hh\:mm\:ss\.ff")}";
+                time_display_label.Text = $"{current.ToString(@"mm\:ss\.ff")} / {total.ToString(@"mm\:ss\.ff")}";
 
             }
         }
@@ -160,6 +170,22 @@ namespace プロコン部チーム_0622_TEST
         {
             trackbar_mouseup = true;
         }
+
+
+        private void Form3_KeyDown(object sender, KeyEventArgs e)
+        {
+            // Spaceキーが押されたかを判定
+            if (e.KeyCode == Keys.C)
+            {
+                Properties.Settings.Default.cut_start_time = TimeSpan.FromSeconds(axWindowsMediaPlayer1.Ctlcontrols.currentPosition).ToString();
+                cut_start_time_display_label.Text = $"カット開始時間: {Properties.Settings.Default.cut_start_time}";
+            }
+
+            if (e.KeyCode == Keys.V)
+            {
+                Properties.Settings.Default.cut_end_time = TimeSpan.FromSeconds(axWindowsMediaPlayer1.Ctlcontrols.currentPosition).ToString();
+                cut_end_time_display_label.Text = $"カット終了時間: {Properties.Settings.Default.cut_end_time}";
+            }
+        }
     }
 }
-
